+++ inc/bot/commands/PingBotCommand.hpp
@@ -6,7 +6,7 @@
 /*   By: ysabik <ysabik@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/26 12:24:30 by ysabik            #+#    #+#             */
-/*   Updated: 2024/05/28 01:06:46 by ysabik           ###   ########.fr       */
+/*   Updated: 2024/05/28 10:45:22 by ysabik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,7 @@ class PingBotCommand : public Executor {
 		bool	exec(Bot *bot, std::string label, std::string prefix,
 					std::vector<std::string> args, std::string sender,
 					std::string target);
+		bool	isElementary();
 };
 
 #endif