+++ src/bot/commands/QuoteExecutor.cpp
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   PingExecutor.cpp                                   :+:      :+:    :+:   */
+/*   QuoteExecutor.cpp                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ysabik <ysabik@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/27 21:59:56 by ysabik            #+#    #+#             */
-/*   Updated: 2024/05/28 10:57:02 by ysabik           ###   ########.fr       */
+/*   Updated: 2024/05/28 10:57:28 by ysabik           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,28 +17,28 @@
 /* ************************************************************************** */
 
 
-PingExecutor::PingExecutor() : Executor() {
+QuoteExecutor::QuoteExecutor() : Executor() {
 }
 
 
-PingExecutor::PingExecutor(std::string name, std::vector<std::string> aliases)
+QuoteExecutor::QuoteExecutor(std::string name, std::vector<std::string> aliases)
 	: Executor(name, aliases) {
 }
 
 
-PingExecutor::PingExecutor(const PingExecutor &command) : Executor() {
+QuoteExecutor::QuoteExecutor(const QuoteExecutor &command) : Executor() {
 	*this = command;
 }
 
 
-PingExecutor &PingExecutor::operator=(const PingExecutor &command) {
+QuoteExecutor &QuoteExecutor::operator=(const QuoteExecutor &command) {
 	if (this != &command)
 		Executor::operator=(command);
 	return *this;
 }
 
 
-PingExecutor::~PingExecutor() {
+QuoteExecutor::~QuoteExecutor() {
 }
 
 
@@ -47,24 +47,36 @@ PingExecutor::~PingExecutor() {
 /* ************************************************************************** */
 
 
-std::string	PingExecutor::getDesc() {
-	return "Ping. Sends a pong.";
+std::string	QuoteExecutor::getDesc() {
+	return "Get a random quote from Quotable.io.";
 }
 
 
-std::string	PingExecutor::getUsage() {
-	return name + " [<whatever>]";
+std::string	QuoteExecutor::getUsage() {
+	return name;
 }
 
 
-bool	PingExecutor::exec(Bot *bot, std::string label, std::string prefix,
+bool	QuoteExecutor::exec(Bot *bot, std::string label, std::string prefix,
 			std::vector<std::string> args, std::string sender, std::string target) {
 	(void) label;
 	(void) prefix;
+	(void) args;
+	(void) sender;
 
-	if (args.size() == 0)
-		bot->send(target, "Pong " + sender + "!");
-	else
-		bot->send(target, "Pong " + args[0] + "!");
+	std::string	quote = curlRequest("https://api.quotable.io/random");
+
+	if (quote.empty())
+		bot->send(target, "I'm sorry, I can't tell you a quote right now.");
+	else {
+		std::string	author = quote.substr(quote.find("\"author\":\"") + 10);
+		author = author.substr(0, author.find("\""));
+
+		quote = quote.substr(quote.find("\"content\":\"") + 11);
+		quote = quote.substr(0, quote.find("\""));
+
+		bot->send(target, quote);
+		bot->send(target, " - " + author);
+	}
 	return true;
 }